sources restored back from revision 2317
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@2432 3895a583-e809-0410-9bb6-aa9463026377

+++ GEM/num_methods.h
@@ -24,7 +24,7 @@ double LagranInterp(float *y, float *x, double *d, float yoi,
 double LagranInterp(float *y, float *x, float *d, float yoi,
                     float xoi, int M, int N, int pp );
 double LagranInterp(double *y, double *x, double *d, double yoi,
-                    double xoi, long int M, long int N, long int pp );
+                    double xoi, int M, int N, int pp );
 #endif   // _num_methods_h_
 
 //-----------------------End of num_methods.h--------------------------
